backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
celery>=5.4,<6.0
redis>=5.0,<6.0
gunicorn>=22.0,<23.0
psycopg2-binary>=2.9,<3.0
django-cors-headers>=4.3,<5.0
django-environ>=0.11,<1.0
boto3>=1.34,<2.0
django-storages>=1.14,<2.0
